debug

+++ src/main/java/dev/hackaton/agua/domain/model/EventRequest.java
@@ -5,7 +5,6 @@ import lombok.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -17,13 +16,13 @@ public class EventRequest implements Serializable {
 //    private static final long serialVersionUID = 1L;
 
     @NotNull(message = "A valid activity must be provided!")
-    private Integer activity;
+    private Long activity;
 
     @NotBlank(message = "Time is mandatory")
     private String time;
 
     @NotBlank(message = "A valid user must be provided!")
-    private Integer user;
+    private Long user;
     private String name;
 
     @NotNull(message = "value for consumption is mandatory")
